Shift + A menu fix
Fixed the bug where the modal wouldn't work in the shift + A add menu in edit mode by changing the layout operator context to invoke default, I'm not sure if this affects the entire menu or just the buttons I add after. I did limited testing on it and didn't notice any difference so until proven otherwise I will assume this is an okay fix.